Remove pickling and scanning before connection.
Looking through the bleak code it seems that all the connect function does when given a BLEDevice object is pull out the address, so there is no advantage to scanning and getting a full BLEObject first.
Instead we can just provide the address as configured when connecting. This may also help with issues on Mac OS 12 where scanning no longer seems possible.
As we are not creating a BLEDevice then we can also remove the pickling functionality.