Removed user_id from config
There is no need to store user_id in the configuration file because we
can read from controller and modify only one byte. Storing
in a configuration file may also cause some issues when someone
incorrectly set the byte responsible for the ability to
move the desk.
I also noticed that BASE_HEIGHT offset is different for my desk so
I added a read from the controller to be fully synchronized.